% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruxjars-package.R
\docType{package}
\name{cruxjars}
\alias{cruxjars}
\alias{cruxjars-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/cruxjars}
\item BugReports: \url{https://gitlab.com/hrbrmstr/cruxjars/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
