% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruxjars-package.R
\docType{package}
\name{cruxjars}
\alias{cruxjars}
\alias{cruxjars-package}
\title{Java Archive Wrapper Supporting the 'crux' Package}
\description{
Contents of the 'crux' & supporting Java archives
\url{https://github.com/chimbori/crux}.Version number reflects the version
number of the included 'JAR' file.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/cruxjars}
\item BugReports: \url{https://gitlab.com/hrbrmstr/cruxjars/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
